producer/requirements.txt
psutil
GPUtil
kafka-python